KSR/requirements.txt
absl-py==0.13.0
anyio==3.3.0
argon2-cffi==20.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach==4.0.0
boto3==1.18.25
botocore==1.21.25
brotlipy==0.7.0
cachetools==4.2.2
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
conda==4.9.2
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cycler==0.10.0
debugpy==1.4.1
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.1.0
entrypoints==0.3
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
gevent==21.8.0
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-auth==1.35.0
google-auth-oauthlib==0.4.5
greenlet==1.1.1
grpcio==1.39.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
inotify-simple==1.2.1
ipykernel==6.2.0
ipython @ file:///tmp/build/80754af9/ipython_1610724799192/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///tmp/build/80754af9/jedi_1606932564285/work
Jinja2 @ file:///home/linux1/recipes/ci/jinja2_1610990516718/work
jmespath==0.10.0
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.1
kiwisolver==1.3.1
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.1
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.1.48
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1609786786540/work
pkginfo==1.7.0
prometheus-client==0.11.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.17.3
psutil @ file:///tmp/build/80754af9/psutil_1598370257551/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
PyNaCl==1.4.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-etcd==0.4.5
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyYAML==5.3.1
pyzmq==22.2.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.7.2
ruamel-yaml==0.15.87
s3transfer==0.5.0
sagemaker-training==3.9.2
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1605205327372/work
sniffio==1.2.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1607965878077/work
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.11.1
testpath==0.5.0
threadpoolctl==2.2.0
torch==1.7.1
torchelastic==0.2.1
torchvision==0.8.2
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
